backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
numpy==1.24.3
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.0.0
scikit-learn==1.3.1
